refactor patches
