+++ JG/internal/JG/CloudUpdateFix.hpp
@@ -1,0 +1,4 @@
+#pragma once
+namespace CloudUpdateFix {
+	void Install();
+}
