+++ JG/internal/JG/CustomHUDShake.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace CustomHUDShake {
+	void Init();
+	void Reset();
+	void Update();
+	float Get(uint8_t modId);
+	void Set(uint8_t modId, float power);
+}
