+++ JG/internal/JG/DisabledArrowKeys.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace DisabledArrowKeys {
+	void Install();
+	void Reset();
+	void Toggle(bool toggle);
+}
