+++ JG/internal/JG/DisabledLevelUp.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace DisabledLevelUp {
+	extern bool isShowLevelUp;
+	void Install();
+	void Reset();
+}
