+++ JG/internal/JG/DisabledMuzzleFlashLights.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace DisabledMuzzleFlashLights {
+	void Install();
+	void Reset();
+	uint32_t SetMode(uint32_t mode);
+}
