+++ JG/internal/JG/ExtraMarkerIcons.hpp
@@ -6,5 +6,5 @@ namespace ExtraMarkerIcons {
 	extern std::unordered_map<uint32_t, std::string> customMapMarkerIcons;
 	void SetMapMarkerIcon(TESObjectREFR* marker, char* iconPath);
 	char* __fastcall GetMapMarker(TESObjectREFR* thisObj, uint16_t mapMarkerType);
-	void Hook();
+	void Install();
 };
