+++ JG/internal/JG/ExtraMiscStats.hpp
@@ -6,9 +6,10 @@ namespace ExtraMiscStats {
 	extern std::unordered_map<std::string, int> miscStatMap;
 	extern std::unordered_set<std::string> availableMiscStats;
 	extern uint8_t recalculateStatFilters;
-	void ResetMap();
+	void Reset();
 	void UpdateMiscStatList(const char* name, int value);
 	bool __cdecl ShouldHideStat(uint32_t* id);
 
-	void Hook();
+	void Install();
+	void Update();
 }
