+++ JG/internal/JG/ExtraReputationIcons.hpp
@@ -3,5 +3,5 @@
 
 namespace ExtraReputationIcons {
 	extern std::unordered_map <uint32_t, std::vector<const char*>> factionRepIcons;
-	void Hook();
+	void Install();
 }
