+++ JG/internal/JG/JohnnyFixes.hpp
@@ -1,0 +1,5 @@
+#pragma once
+namespace JohnnyFixes {
+	void Install();
+	void ClearPlayerFurniture();
+}
