+++ JG/internal/JG/JohnnyRadios.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace JohnnyRadios {
+	void Reset();
+	void Update();
+	bool IsAvailable(uint32_t formID);
+}
