+++ JG/internal/JG/MediaLocationControllerOverride.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include "GameForms.h"
+
+namespace MediaLocationControllerOverride {
+	void Install();
+	void Reset();
+	void Set(MediaLocationController* ctrl);
+}
