+++ JG/internal/JG/QuestObjectiveDisplayFix.cpp
@@ -1,0 +1,7 @@
+#include "QuestObjectiveDisplayFix.hpp"
+namespace QuestObjectiveDisplayFix {
+	void Install()
+	{
+		QuestObjectiveDisplayHook<0x05A5E70>();
+	}
+}
