+++ JG/nvse_plugin_example.vcxproj
@@ -180,17 +180,28 @@ copy "$(TargetDir)$(TargetName).pdb" "C:\Symbols\$(TargetName).pdb"
     <ClCompile Include="internal\Game\Gamebryo\NiMemory.cpp" />
     <ClCompile Include="internal\Game\Gamebryo\NiObject.cpp" />
     <ClCompile Include="internal\JG\BarterFilter.cpp" />
+    <ClCompile Include="internal\JG\CameraOverride.cpp" />
+    <ClCompile Include="internal\JG\CloudUpdateFix.cpp" />
     <ClCompile Include="internal\JG\CustomCameraShake.cpp" />
+    <ClCompile Include="internal\JG\CustomHUDShake.cpp" />
     <ClCompile Include="internal\JG\DialogueResponseOverride.cpp" />
+    <ClCompile Include="internal\JG\DisabledArrowKeys.cpp" />
+    <ClCompile Include="internal\JG\DisabledLevelUp.cpp" />
+    <ClCompile Include="internal\JG\DisabledMuzzleFlashLights.cpp" />
+    <ClCompile Include="internal\JG\DisabledSaves.cpp" />
     <ClCompile Include="internal\JG\EditorIDRestoration.cpp" />
     <ClCompile Include="internal\JG\ExtraMarkerIcons.cpp" />
     <ClCompile Include="internal\JG\ExtraMiscStats.cpp" />
     <ClCompile Include="internal\JG\ExtraReputationIcons.cpp" />
     <ClCompile Include="internal\JG\ExtraUISounds.cpp" />
+    <ClCompile Include="internal\JG\JohnnyFixes.cpp" />
     <ClCompile Include="internal\JG\JohnnyPatches.cpp" />
     <ClCompile Include="internal\JG\JohnnyGameSettings.cpp" />
     <ClCompile Include="internal\JG\JohnnyExtraData.cpp" />
+    <ClCompile Include="internal\JG\JohnnyRadios.cpp" />
+    <ClCompile Include="internal\JG\MediaLocationControllerOverride.cpp" />
     <ClCompile Include="internal\JG\NPCAccuracy.cpp" />
+    <ClCompile Include="internal\JG\QuestObjectiveDisplayFix.cpp" />
     <ClCompile Include="internal\JG\RSMBarberHook.cpp" />
     <ClCompile Include="internal\JIP\JIPFixes.cpp" />
     <ClCompile Include="internal\md5\md5.cpp" />
@@ -321,18 +332,29 @@ copy "$(TargetDir)$(TargetName).pdb" "C:\Symbols\$(TargetName).pdb"
     <ClInclude Include="internal\Game\Gamebryo\NiTStringTemplateMap.hpp" />
     <ClInclude Include="internal\havok.h" />
     <ClInclude Include="internal\JG\BarterFilter.hpp" />
+    <ClInclude Include="internal\JG\CameraOverride.hpp" />
+    <ClInclude Include="internal\JG\CloudUpdateFix.hpp" />
     <ClInclude Include="internal\JG\CustomCameraShake.hpp" />
+    <ClInclude Include="internal\JG\CustomHUDShake.hpp" />
     <ClInclude Include="internal\JG\DialogueResponseOverride.hpp" />
+    <ClInclude Include="internal\JG\DisabledArrowKeys.hpp" />
+    <ClInclude Include="internal\JG\DisabledLevelUp.hpp" />
+    <ClInclude Include="internal\JG\DisabledMuzzleFlashLights.hpp" />
+    <ClInclude Include="internal\JG\DisabledSaves.hpp" />
     <ClInclude Include="internal\JG\EditorIDRestoration.hpp" />
     <ClInclude Include="internal\JG\ExtraMarkerIcons.hpp" />
     <ClInclude Include="internal\JG\ExtraMiscStats.hpp" />
     <ClInclude Include="internal\JG\ExtraReputationIcons.hpp" />
     <ClInclude Include="internal\JG\ExtraUISounds.hpp" />
     <ClInclude Include="internal\JG\JGSetList.hpp" />
+    <ClInclude Include="internal\JG\JohnnyFixes.hpp" />
     <ClInclude Include="internal\JG\JohnnyPatches.hpp" />
     <ClInclude Include="internal\JG\JohnnyGameSettings.hpp" />
     <ClInclude Include="internal\JG\JohnnyExtraData.hpp" />
+    <ClInclude Include="internal\JG\JohnnyRadios.hpp" />
+    <ClInclude Include="internal\JG\MediaLocationControllerOverride.hpp" />
     <ClInclude Include="internal\JG\NPCAccuracy.hpp" />
+    <ClInclude Include="internal\JG\QuestObjectiveDisplayFix.hpp" />
     <ClInclude Include="internal\JG\RSMBarberHook.hpp" />
     <ClInclude Include="internal\JIP\JIPFixes.hpp" />
     <ClInclude Include="internal\md5\md5.h" />
